ipc4: Expose ipc4_update_buffer_format()
It will be needed by other modules as well.
