+++ src/include/ipc4/base-config.h
@@ -25,6 +25,7 @@
 #define __SOF_IPC4_BASE_CONFIG_H__
 
 #include <stdint.h>
+#include <sof/compiler_attributes.h>
 
 enum ipc4_sampling_frequency {
 	IPC4_FS_8000HZ = 8000,
@@ -240,5 +241,8 @@ union ipc4_cfg_param_id_data {
 struct sof_ipc_stream_params;
 void ipc4_base_module_cfg_to_stream_params(const struct ipc4_base_module_cfg *base_cfg,
 					   struct sof_ipc_stream_params *params);
+struct comp_buffer;
+void ipc4_update_buffer_format(struct comp_buffer __sparse_cache *buf_c,
+			       const struct ipc4_audio_format *fmt);
 
 #endif
